Refactor utils to utils_legacy. This will be deleted completly in the future when migration is complete
