+++ familytree/add_person.py
@@ -15,7 +15,7 @@ from PySide6.QtWidgets import (
     QVBoxLayout,
     QWidget,
 )
-from utils import ProtoUtility
+from utils_legacy import ProtoUtility
 
 # Get a logger instance for this module
 logger = logging.getLogger(__name__)
